Revert "Move migrate package"
This reverts commit 62349320c6e9481a09300ef2f653e6a1d4b41372.
