+++ storage/postgres/migrate/copied.go
@@ -5,9 +5,7 @@ import (
 	"crypto/cipher"
 	"encoding/json"
 	"fmt"
-
 	"github.com/MicahParks/jwkset"
-
 	"github.com/MicahParks/magiclinksdev/storage"
 )
